vm: move register logic into register.rs
