+++ compiler/hash-vm/src/lib.rs
@@ -7,6 +7,8 @@ mod heap;
 mod stack;
 
 pub mod bytecode;
+pub mod register;
+
 pub mod bytecode_builder;
 pub mod error;
 pub mod vm;
